refactor : Permission은 싱글 모드에서도 사용해야 하므로 feature/battle -> feature/running 으로 이동
